gen-ai-course/14_ai_projects/customer_support_engine/pyproject.toml
[project]
name = "customer-support-engine"
version = "0.1.0"
description = "Stateful multi-agent customer support system using LangGraph + Azure AI"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.1.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "azure-search-documents>=11.4.0",
    "azure-cosmos>=4.6.0",
    "azure-identity>=1.15.0",
    "azure-monitor-opentelemetry>=1.4.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
